Dedupe object flag script
